Fix black screen for large UI trees
Adjust the UI element drawing layer input to the shape drawing
auxiliary by normalizing against the maximum layer number.
Normalization is required for screen space.

+++ src/draw/shapes.c
@@ -35,7 +35,7 @@ void shapesBegin()
 }
 
 void shapesGenerateBox(float lower_x, float upper_x, float lower_y, float upper_y,
-                       int layer, struct RGBA color,
+                       float layer, struct RGBA color,
                        unsigned int *VAO, unsigned int *VBO)
 {
     float vertices[] =
