+++ src/draw/shapes.h
@@ -2,6 +2,6 @@
 
 void shapesBegin();
 void shapesGenerateBox(float upper_x, float lower_x, float upper_y, float lower_y,
-                       int layer, struct RGBA color,
+                       float layer, struct RGBA color,
                        unsigned int *VAO, unsigned int *VBO);
 void shapesDrawBox(unsigned int *VAO);
